dev/python_frontend/requirements.txt
numpy==1.26.4
matplotlib==3.8.3
scipy==1.11.1
opencv-python==4.9.0.80
pyserial==3.5